+++ Libs/Utils/HashMD5.h
@@ -17,7 +17,7 @@ class CHashMD5
 {
 protected:
 	/* typedef a 32 bit type */
-	typedef unsigned long int UINT4;
+	typedef unsigned int UINT4;
 
 	/* Data structure for MD5 (Message Digest) computation */
 	typedef struct {
@@ -33,7 +33,7 @@ public:
 	CHashMD5(void);
 	~CHashMD5(void);
 	
-	void AddData(BYTE * inBuf, unsigned int inLen);
+	void AddData(BYTE * inBuf, size_t inLen);
 	void MD5Final(void);
 	CEasyString GetHashCodeString(void);
 	void GetHashCode(BYTE * pBuff);
